refactor: Extract meshData from App.js
