+++ src/components/three/DirectionalBox.js
@@ -53,13 +53,13 @@ const DirectionalBox = forwardRef((props, ref) => {
                     </Html>
                 )}
             </mesh>
-            <mesh position={[0, (boxDimensions[1] / 2) + 0.125, 0]}>
+            <mesh position={[0, (transform.boxDimensions[1] / 2) + 0.125, 0]}>
                 <cylinderGeometry args={[0.05, 0.05, 0.5, 32]}/>
-                <meshStandardMaterial color={(hovered) ? "#c02040" : "yellow"}/>
+                <meshStandardMaterial color={(interaction.hovered) ? "#c02040" : "yellow"}/>
             </mesh>
-            <mesh position={[0, (boxDimensions[1] / 2) + 0.625, 0]}>
+            <mesh position={[0, (transform.boxDimensions[1] / 2) + 0.625, 0]}>
                 <coneGeometry args={[0.125, 0.5, 4, 1]}/>
-                <meshStandardMaterial color={(hovered) ? "#c02040" : "yellow"}/>
+                <meshStandardMaterial color={(interaction.hovered) ? "#c02040" : "yellow"}/>
             </mesh>
         </group>
     )
