+++ src/constants/meshData.js
@@ -1,0 +1,5 @@
+ export const meshData = [
+    { id: 55, position: [-5, 0, 0], rotation: [-Math.PI / 4, 0, 0], size: [1, 0.5, 4], description: "Antenna Object 55"},
+    { id: 66, position: [0, 0, 0], rotation: [0, 0, 0], size: [1, 0.5, 4], description: "Antenna Object 66"},
+    { id: 77, position: [5, 0, 0], rotation: [Math.PI/4, 0, 0], size: [1, 0.5, 4], description: "Antenna Object 77"},
+];
